requirements.txt
mcp
langchain
langchain-openai
langchain-community
chromadb
pydantic
pandas
sqlalchemy
rich
pyyaml
numpy
scipy